+++ NEWS.md
@@ -3,6 +3,7 @@
 - Added the capacity to make a gap between groups
 - Added a function to generate additional noise variables
 - Added a function to make predictors correlated
+- Added normal sampling as an alternative to uniform and gridded
 
 # sillysplines 1.0.0